Refactored code : moved comment feature to new files ( CommentController, commentService and CommentRepository )
So that the test cases can pass and MVC architrure is followed.
