+++ src/main/java/ImageHoster/repository/ImageRepository.java
@@ -116,18 +116,6 @@ public class ImageRepository {
             transaction.rollback();
         }
     }
-    // Add Comment to db
-    public void addComment (Comment comment) {
-        EntityManager em = emf.createEntityManager();
-        EntityTransaction transaction = em.getTransaction();
 
-        try {
-            transaction.begin();
-            em.persist(comment);
-            transaction.commit();
-        } catch (Exception e) {
-            transaction.rollback();
-        }
-    }
 
 }
