+++ src/main/java/ImageHoster/service/CommentService.java
@@ -1,0 +1,17 @@
+package ImageHoster.service;
+
+import ImageHoster.model.Comment;
+import ImageHoster.repository.CommentRepository;
+import ImageHoster.repository.ImageRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CommentService {
+    @Autowired
+    private CommentRepository commentRepository;
+
+    public void addComment(Comment comment) {
+        commentRepository.addComment(comment);
+    }
+}
