+++ src/test/java/ImageHoster/controller/HomeControllerTest.java
@@ -1,4 +1,4 @@
-/*
+
 package ImageHoster.controller;
 
 import ImageHoster.service.ImageService;
@@ -36,4 +36,4 @@ public class HomeControllerTest {
                 .andExpect(content().string(containsString("Image Hoster")));
     }
 }
-*/
+
