+++ src/test/java/ImageHoster/controller/ImageControllerTest.java
@@ -1,4 +1,4 @@
-/*
+
 package ImageHoster.controller;
 
 import ImageHoster.model.Image;
@@ -310,4 +310,3 @@ public class ImageControllerTest {
     }
 }
 
-*/
